[gaim-migrate @ 16492] Fix a casting warning on 64-bit machines
--HG--
extra : convert_revision : 37e900d75721580a594519905ee3b16fabdba651

+++ src/protocols/oscar/oscar.c
@@ -1629,7 +1629,7 @@ gaim_handle_redirect(OscarData *od, FlapConnection *conn, FlapFrame *fr, ...)
 	separator = strchr(redir->ip, ':');
 	if (separator != NULL)
 	{
-		host = g_strndup(redir->ip, (int)separator - (int)redir->ip);
+		host = g_strndup(redir->ip, separator - redir->ip);
 		port = atoi(separator + 1);
 	}
 	else
